To Long for money amount type. Added some other domain/business objects and queries

+++ app/src/main/java/it/gr85/android/apps/em/adapters/android/room/ARTransactionDao.kt
@@ -46,8 +46,8 @@ interface ARTransactionDao {
         categoryId: String?,
         type: String?,
         descriptionQuery: String?,
-        minAmount: Float?,
-        maxAmount: Float?
+        minAmount: Long?,
+        maxAmount: Long?
     ): List<ARTransactionEntity>
 
     @Insert
